+++ CHANGELOG.md
@@ -13,6 +13,8 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
 - `LightGBM` model now supports native categorical feature handling as described
    [here](https://lightgbm.readthedocs.io/en/latest/Features.html#optimal-split-for-categorical-features).
    [#1585](https://github.com/unit8co/darts/pull/1585) by [Rijk van der Meulen](https://github.com/rijkvandermeulen)
+- Fix issue when resuming training from checkpoint Torch models 
+   [#1495](https://github.com/unit8co/darts/pull/1495) by [Rijk van der Meulen](https://github.com/rijkvandermeulen)
 
 ## [0.23.1](https://github.com/unit8co/darts/tree/0.23.1) (2023-01-12)
 Patch release